Little logic error in the HLE ioman code that arcum spotted.  (It's in code that isn't reachable unless there's a bug in fd allocation so it won't affect anything.)
git-svn-id: http://pcsx2.googlecode.com/svn/trunk@3129 96395faa-99c1-11dd-bbfe-3dabce05a288

+++ pcsx2/IopBios.cpp
@@ -248,7 +248,7 @@ namespace ioman {
 			else
 			{
 				v0 = allocfd(file);
-				if (v0 < 0)
+				if ((s32)v0 < 0)
 					file->close();
 			}
